Move: BrandProduct to product doamin
