+++ src/main/java/org/example/sivillage/product/domain/BrandProduct.java
@@ -1,4 +1,4 @@
-package org.example.sivillage.brand.domain;
+package org.example.sivillage.product.domain;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
