+++ src/main/java/org/example/sivillage/product/dto/in/CreateProductRequestDto.java
@@ -3,7 +3,7 @@ package org.example.sivillage.product.dto.in;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.example.sivillage.brand.domain.BrandProduct;
+import org.example.sivillage.product.domain.BrandProduct;
 import org.example.sivillage.global.common.UuidGenerator;
 import org.example.sivillage.product.domain.Product;
 import org.example.sivillage.product.vo.in.CreateProductRequestVo;
