+++ src/main/java/org/example/sivillage/product/infrastructure/BrandProductRepository.java
@@ -1,6 +1,6 @@
 package org.example.sivillage.product.infrastructure;
 
-import org.example.sivillage.brand.domain.BrandProduct;
+import org.example.sivillage.product.domain.BrandProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
